// src/app/shared/models/faction.enum.ts
// src/app/shared/models/colors.enum.ts
export enum Faction {
    Terrans = 'Terrans',
    Lantids = 'Lantids',
    Xenos = 'Xenos',
    Gleens = 'Gleens',
    Taklons = 'Taklons',
    Ambas = 'Ambas',
    HadschHallas = 'Hadsch Hallas',
    Ivits = 'Ivits',
    Geodens = 'Geodens',
    BalTaks = "Bal T'aks",
    Firaks = 'Firaks',
    Bescods = 'Bescods',
    Nevlas = 'Nevlas',
    Itars = 'Itars',
}